backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0          # async HTTP client for fetchers
python-dotenv>=1.0.0   # load .env.local / .env
pydantic>=2.7.0        # data models / validation
